requirements.txt
protobuf==6.33.0
python-dotenv==1.1.1
